Move `Link` export
