+++ packages/next-intl/link.d.ts
@@ -1,0 +1,1 @@
+export {default} from './dist/src/link';
