+++ packages/next-intl/src/link/index.tsx
@@ -1,0 +1,1 @@
+export {default} from './Link';
